# api/openapi.yaml
openapi: 3.0.3
info:
  title: asyou Management API
  version: v1
servers:
  - url: /api/v1
paths:
...
  /api-keys:
    post:
      summary: Create a new API key
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyCreate'
      responses:
        '201':
          description: created (token returned once)
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
    get:
      summary: List user's API keys
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApiKey'
  /api-keys/{id}:
    delete:
      summary: Revoke an API key
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '204':
          description: revoked
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
  schemas:
...
    ApiKey:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        name:
          type: string
        scopes:
          type: string
        revoked:
          type: boolean
        created_at:
          type: string
          format: date-time
    ApiKeyCreate:
      type: object
      properties:
        name:
          type: string
        scopes:
          type: string